fist combine

+++ src/main/java/ca/team3/laps/service/AdminStaffService.java
@@ -6,7 +6,7 @@ import ca.team3.laps.model.Staff;
 
 public interface AdminStaffService {
     List<Staff> findAllStaff();
-    Staff findStaffById(long id);
+    Staff findStaffById(Integer id);
     void createStaff(Staff staff);
     void modifyStaff(Staff staff);
     void deleteStaff(Staff staff);
